requirements.txt
Flask==1.1.2
requests==2.24.0
botbuilder-core==4.10.1
botbuilder-schema==4.10.1
botframework-connector==4.10.1
APScheduler==3.6.3
kubernetes==11.0.0
holidays==0.10.3
beautifulsoup4==4.9.3
oyaml==1.0
click-shell[readline]==2.0
pytz==2020.1
